requirements.txt
opencv-python
numpy
pillow
pytesseract
easyocr
rapidfuzz
pandas
streamlit
pytest
python-dotenv
